Organized files and changed references to files in others
