+++ resources/assets/js/playerUpdate.js
