+++ resources/views/auth/playerUpdate.blade.php
@@ -54,5 +54,5 @@
 <button onclick="">Logout</button>
 </body>
 <script src="https://code.jquery.com/jquery-3.2.1.js" integrity="sha256-DZAnKJ/6XZ9si04Hgrsxu/8s717jcIzLy3oi35EouyE=" crossorigin="anonymous"></script>
-<script src="/LeagueOfLegendsDisplay/JS/playerUpdate.js"></script>
+<script src="/resources/assets/js/playerUpdate.js"></script>
 </html>
